[Flat] remove kernel functions from operator (#533)
* [Flat] Kernel refactor

* some fixes

* remove static

* clang formatting fixes

* clang formatting fixes

* clang formatting fixes

* clang formatting fixes

* clang formatting fixes

* clang formatting fixes

* temporary fix for ci

* temporary fix for ci

* temporary fix for ci

* temporary fix for ci

* temporary fix for ci

* formatting fix

* formatting fix

* undo temporary fixes for ci

* undo temporary fixes for ci

* Apply suggestions from code review

Remove static
